+++ source/extensions/omni.isaac.lab_tasks/omni/isaac/lab_tasks/direct/allegro_hand/__init__.py
@@ -26,24 +26,3 @@ gym.register(
         "rsl_rl_cfg_entry_point": agents.rsl_rl_ppo_cfg.AllegroHandPPORunnerCfg,
     },
 )
-
-gym.register(
-    id="Isaac-Allegro-Hand-OpenAI-FF-Direct-v0",
-    entry_point="omni.isaac.lab_tasks.direct.inhand_manipulation:InHandManipulationEnv",
-    disable_env_checker=True,
-    kwargs={
-        "env_cfg_entry_point": AllegroHandOpenAIEnvCfg,
-        "rl_games_cfg_entry_point": f"{agents.__name__}:rl_games_ppo_ff_cfg.yaml",
-        "rsl_rl_cfg_entry_point": agents.rsl_rl_ppo_cfg.AllegroHandAsymFFPPORunnerCfg,
-    },
-)
-
-gym.register(
-    id="Isaac-Allegro-Hand-OpenAI-LSTM-Direct-v0",
-    entry_point="omni.isaac.lab_tasks.direct.inhand_manipulation:InHandManipulationEnv",
-    disable_env_checker=True,
-    kwargs={
-        "env_cfg_entry_point": AllegroHandOpenAIEnvCfg,
-        "rl_games_cfg_entry_point": f"{agents.__name__}:rl_games_ppo_lstm_cfg.yaml",
-    },
-)